+++ static/src/javascripts/projects/commercial/modules/ipsosTaggingScript.js
